TST Move fixtures up a level [scipy-dev]
